Added functionality to delete a player from the database.
