+++ src/components/dialogs/DeleteDialog.java
@@ -1,11 +1,15 @@
 package components.dialogs;
 
+import backend.PlayerProvider;
+
 import javax.swing.*;
+import java.sql.SQLException;
 
 /**
  * This is the popup / dialog window for removing players.
  */
 public class DeleteDialog extends InputDialog {
+    private PlayerProvider playerProvider = new PlayerProvider();
 
     /**
      * Create and instantiate the custom dialog window
@@ -29,7 +33,11 @@ public class DeleteDialog extends InputDialog {
         int option = JOptionPane.showOptionDialog(this, inputFields, "Speler verwijderen", JOptionPane.OK_CANCEL_OPTION, JOptionPane.INFORMATION_MESSAGE, null, new Object[]{delete, cancel}, delete);
 
         if (option == JOptionPane.OK_OPTION) {
-            // verwijderen
+            try {
+                playerProvider.deletePlayer(id);
+            } catch (SQLException e) {
+                JOptionPane.showMessageDialog(this, "Er is iets fout gegaan met het verwijderen van de speler. Probeer het opnieuw.", "Foutmelding", JOptionPane.ERROR_MESSAGE);
+            }
         }
     }
 }
